requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
flask
flask-login
apscheduler
gunicorn
openpyxl
